refactor: extract users route in own controller
