+++ src/services/http/controllers/interface.ts
@@ -1,0 +1,6 @@
+import { GenericService } from '../../../util/svc';
+import { Router } from 'express';
+
+export abstract class Controller extends GenericService {
+  router: Router;
+}
