+++ src/services/http/http.test.ts
@@ -2,7 +2,7 @@ import 'isomorphic-fetch';
 import tape from 'tape';
 import sinon from 'sinon';
 import HttpService from '.';
-import { stubCall, stubFetch } from '../../util/testUtils';
+import { newRequest, newResponse, stubCall, stubFetch } from '../../util/testUtils';
 
 const post = {
   type: 'POST',
@@ -33,221 +33,6 @@ const post = {
     modFollowerUser: null,
   },
 };
-
-const newRequest = (params?: any, body?: any, query?: any): any => {
-  return {
-    query: {
-      limit: 10,
-      offset: 0,
-      ...query,
-    },
-    params,
-    body,
-    header: () => null,
-    session: {},
-  };
-};
-
-const newResponse = (): any => {
-  const ret: any = {
-    send: sinon.stub(),
-  };
-
-  ret.status = sinon.stub().returns(ret);
-
-  return ret;
-};
-
-tape('HTTPService - handleGetUser', async t => {
-  const http = new HttpService();
-  const [call, stubs] = stubCall(http);
-  const req = newRequest();
-  const res = newResponse();
-
-  stubs.users.readAll.returns(Promise.resolve([{ username: '0xuser1' }, { username: '0xuser2' }]));
-  call.withArgs('ens', 'fetchNameByAddress').returns(Promise.resolve(''));
-
-  await http.handleGetUser(req, res);
-
-  t.deepEqual(
-    stubs.users.readAll.args[0],
-    [undefined, 0, 10],
-    'should read users db with correct params'
-  );
-
-  t.deepEqual(
-    res.send.args[0],
-    [
-      {
-        payload: [
-          { ens: '', username: '0xuser1' },
-          { ens: '', username: '0xuser2' },
-        ],
-        error: undefined,
-      },
-    ],
-    'should return list of users'
-  );
-
-  t.end();
-});
-
-tape('HTTPService - handleSearchUser', async t => {
-  const http = new HttpService();
-  const [call, stubs] = stubCall(http);
-  const req = newRequest({ query: '2728' });
-  const res = newResponse();
-
-  stubs.users.search.returns(
-    Promise.resolve([{ username: '0xsearchuser1' }, { username: '0xsearchuser2' }])
-  );
-  call.withArgs('ens', 'fetchNameByAddress').returns(Promise.resolve(''));
-
-  await http.handleSearchUser(req, res);
-
-  t.deepEqual(
-    stubs.users.search.args[0],
-    ['2728', undefined, 0, 10],
-    'should search users db with correct params'
-  );
-
-  t.deepEqual(
-    res.send.args[0],
-    [
-      {
-        payload: [
-          { ens: '', username: '0xsearchuser1' },
-          { ens: '', username: '0xsearchuser2' },
-        ],
-        error: undefined,
-      },
-    ],
-    'should return list of users'
-  );
-
-  t.end();
-});
-
-tape('HTTPService - handleGetUserByAddress', async t => {
-  const http = new HttpService();
-  const [call, stubs] = stubCall(http);
-  const req = newRequest({ address: '0xd44a82dD160217d46D754a03C8f841edF06EBE3c' });
-  const res = newResponse();
-
-  stubs.users.findOneByName.returns(
-    Promise.resolve({
-      username: '0xmyaddress',
-    })
-  );
-
-  call.withArgs('ens', 'fetchNameByAddress').returns(Promise.resolve(''));
-
-  await http.handleGetUserByAddress(req, res);
-
-  t.deepEqual(
-    stubs.users.findOneByName.args[0],
-    ['0xd44a82dD160217d46D754a03C8f841edF06EBE3c', undefined],
-    'should get user with correct params'
-  );
-
-  t.deepEqual(
-    res.send.args[0],
-    [
-      {
-        payload: {
-          username: '0xd44a82dD160217d46D754a03C8f841edF06EBE3c',
-          ens: '',
-          address: '0xd44a82dD160217d46D754a03C8f841edF06EBE3c',
-        },
-        error: undefined,
-      },
-    ],
-    'should return user'
-  );
-
-  t.end();
-});
-
-tape('HTTPService - handleAddUser', async t => {
-  const http = new HttpService();
-  const [call, stubs] = stubCall(http);
-  const req = newRequest(null, {
-    account: '0xd44a82dD160217d46D754a03C8f841edF06EBE3c',
-    publicKey: '0xpubkey',
-    proof: '0xproof',
-  });
-  const res = newResponse();
-
-  stubs.users.findOneByName.returns(
-    Promise.resolve({
-      username: '0xmyaddress',
-    })
-  );
-
-  call.withArgs('arbitrum', 'getNonce').returns(Promise.resolve(0));
-  call
-    .withArgs('ens', 'ecrecover')
-    .returns(Promise.resolve('0xd44a82dD160217d46D754a03C8f841edF06EBE3c'));
-  call.withArgs('arbitrum', 'updateFor').returns(Promise.resolve('0xtxhash'));
-
-  await http.handleAddUser(req, res);
-
-  t.deepEqual(
-    call.args[1],
-    [
-      'ens',
-      'ecrecover',
-      '0x87777c0e30f6a3458fd35fa202eec524c2b3b6713bbb3c388d43e9db9040867f',
-      '0xproof',
-    ],
-    'should call ecrecover with hash'
-  );
-
-  t.deepEqual(
-    call.args[2],
-    ['arbitrum', 'updateFor', '0xd44a82dD160217d46D754a03C8f841edF06EBE3c', '0xpubkey', '0xproof'],
-    'should call arbitrum contract with correct params'
-  );
-
-  t.deepEqual(
-    res.send.args[0],
-    [
-      {
-        payload: '0xtxhash',
-        error: undefined,
-      },
-    ],
-    'should return tx hash'
-  );
-
-  await http.handleAddUser(newRequest(null, {}), res);
-
-  await http.handleAddUser(
-    newRequest(null, {
-      account: '0xd44a82dD160217d46D754a03C8f841edF06EBE3c',
-    }),
-    res
-  );
-
-  await http.handleAddUser(
-    newRequest(null, {
-      account: '0xd44a82dD160217d46D754a03C8f841edF06EBE3c',
-      publicKey: '0xpubkey',
-    }),
-    res
-  );
-
-  t.equal(res.status.args[0][0], 400, 'should return error code');
-  t.equal(res.send.args[1][0].payload, 'invalid account', 'should return error message');
-
-  t.equal(res.status.args[1][0], 400, 'should return error code');
-  t.equal(res.send.args[2][0].payload, 'invalid publicKey', 'should return error message');
-
-  t.equal(res.status.args[2][0], 400, 'should return error code');
-  t.equal(res.send.args[3][0].payload, 'invalid proof', 'should return error message');
-
-  t.end();
-});
 
 tape('HTTPService - handleGetReplies', async t => {
   const http = new HttpService();
