+++ src/services/http/index.ts
@@ -162,98 +162,6 @@ export default class HttpService extends GenericService {
     };
   }
 
-  handleGetUser = async (req: Request, res: Response) => {
-    const limit = req.query.limit && Number(req.query.limit);
-    const offset = req.query.offset && Number(req.query.offset);
-    const context = req.header('x-contextual-name') || undefined;
-    const usersDB = await this.call('db', 'getUsers');
-    const users = await usersDB.readAll(context, offset, limit);
-
-    const result = [];
-
-    for (let user of users) {
-      const ens = await this.call('ens', 'fetchNameByAddress', user.username);
-      result.push({ ens, ...user });
-    }
-
-    res.send(makeResponse(result));
-  };
-
-  handleSearchUser = async (req: Request, res: Response) => {
-    const limit = req.query.limit && Number(req.query.limit);
-    const offset = req.query.offset && Number(req.query.offset);
-    const query = req.params.query;
-    const context = req.header('x-contextual-name') || undefined;
-    const usersDB = await this.call('db', 'getUsers');
-    const users = await usersDB.search(query || '', context, offset, limit);
-    const result = [];
-
-    for (let user of users) {
-      const ens = await this.call('ens', 'fetchNameByAddress', user.username);
-      result.push({ ens, ...user });
-    }
-
-    res.send(makeResponse(result));
-  };
-
-  handleGetUserByAddress = async (req: Request, res: Response) => {
-    const usersDB = await this.call('db', 'getUsers');
-
-    let address = req.params.address;
-
-    try {
-      address = Web3.utils.toChecksumAddress(address);
-    } catch (e) {}
-
-    if (!Web3.utils.isAddress(address)) {
-      address = await this.call('ens', 'fetchAddressByName', address);
-    }
-
-    const context = req.header('x-contextual-name') || undefined;
-    const user = await usersDB.findOneByName(address, context);
-    const ens = await this.call('ens', 'fetchNameByAddress', address);
-    res.send(
-      makeResponse({
-        ...user,
-        ens: ens,
-        address: address,
-        username: address,
-      })
-    );
-  };
-
-  handleAddUser = async (req: Request, res: Response) => {
-    const { account, publicKey, proof } = req.body;
-
-    if (!account || !Web3.utils.isAddress(account)) {
-      res.status(400).send(makeResponse('invalid account'));
-      return;
-    }
-
-    if (!publicKey) {
-      res.status(400).send(makeResponse('invalid publicKey'));
-      return;
-    }
-
-    if (!proof) {
-      res.status(400).send(makeResponse('invalid proof'));
-      return;
-    }
-
-    const pubkeyBytes = Web3.utils.utf8ToHex(publicKey);
-    const nonce = await this.call('arbitrum', 'getNonce', account);
-    const hash = Web3.utils.keccak256(Web3.utils.encodePacked(account, pubkeyBytes, nonce)!);
-    const recover = await this.call('ens', 'ecrecover', hash, proof);
-
-    if (recover !== account) {
-      throw new Error('invalid signature');
-    }
-
-    const tx = await this.call('arbitrum', 'updateFor', account, publicKey, proof);
-
-    res.send(makeResponse(tx));
-  };
-
   handleGetReplies = async (req: Request, res: Response) => {
     const limit = req.query.limit && Number(req.query.limit);
     const offset = req.query.offset && Number(req.query.offset);
@@ -615,17 +523,20 @@ export default class HttpService extends GenericService {
     res.send(makeResponse('ok'));
   };
 
+  initControllers() {
+    ['users'].forEach(controller => {
+      this.app.use('/', this.get(`${controller}Controller`, 'router'));
+    });
+  }
+
   addRoutes() {
+    this.initControllers();
     this.app.get(
       '/healthcheck',
       this.wrapHandler(async (req, res) => {
         res.send(makeResponse('ok'));
       })
     );
-    this.app.get('/v1/users', this.wrapHandler(this.handleGetUser));
-    this.app.get('/v1/users/search/:query?', this.wrapHandler(this.handleSearchUser));
-    this.app.get('/v1/users/:address', this.wrapHandler(this.handleGetUserByAddress));
-    this.app.post('/v1/users', jsonParser, this.wrapHandler(this.handleAddUser));
     this.app.get('/v1/replies', this.wrapHandler(this.handleGetReplies));
     this.app.get('/v1/posts', this.wrapHandler(this.handleGetPosts));
     this.app.get('/v1/tags/:tagName', this.wrapHandler(this.handleGetPostsByTag));
